User register dispatching changed
As is, if a session has an email defined, that email is looked up in the database, and the associated user is stashed. If not found, the user will be prompted to choose a username, which will make a new user + email record in the database.

This will probably be changed somewhat, so that the user can "logout" (destroy the email in their session) without making a username.