added an auto-increment pk `id` to player
This change was mostly out of the realisation that the player_no gives away a player's role and cannot under any circumstance be communicated with players, so putting the PK in an input field is not possible. To get forms to work without the hassle of looking up players by their alias (which are not strictly unique -- but that's another issue), some other unique identifier is required.